Move statsd package
No functional changes
