+++ cmd/tester/main.go
@@ -9,7 +9,7 @@ import (
 	"time"
 
 	"github.com/atlassian/gostatsd/pkg/fakesocket"
-	"github.com/atlassian/gostatsd/statsd"
+	"github.com/atlassian/gostatsd/pkg/statsd"
 
 	log "github.com/Sirupsen/logrus"
 	"github.com/spf13/pflag"
