+++ pkg/statsd/doc.go
