+++ pkg/statsd/lexer.go
