+++ pkg/statsd/lexer_fuzz.go
