+++ pkg/statsd/lexer_test.go
